requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
dnspython==1.16.0
eventlet==0.31.1
greenlet==1.1.0
httpie==2.4.0
idna==2.10
importlib-metadata==4.6.1
kombu==5.1.0
mysqlclient==2.0.3
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
